+++ CHANGELOG.md
@@ -161,6 +161,7 @@ for label, interval_data in results.groupby("interval_labels"):
 - Warn on no-operation restrictions #1586
 - Improved efficiency for writing multiple objects to analysis file #1594
 - Pin `scipy<1.13` for `spikeinterface==0.99.1` compatibility #1612
+- Active management of memory resources for open NWBs #1606
 
 ### Pipelines